Changed how NASM header files are generated
A hack to workaround make. When genereating the files, the bash script
c_to_nasm.sh will check if the C header is newer than the NASM header,
which means that the C header has been changed since the last make. Only
then will the script update the NASM headers.