Rewrote mkfs and inode.h to make a new filesystem
This rewrite was "necessary" because we want to have /dev, which would
contain special files, and which we therefore would need to write
specially, or mount as a new separate filesystem.

The structure with superblock|inodes|blocks is very simple, and similar
to very basic UFS/ext2.

It doesn't compile - the call to open() needs one more argument in mkfs,
and the kernel code hasn't been modified to keep up with the changes in
inode.h.