Fixed two page frame allocation bugs
First, the mmap[i] was incorrectly checked to see whether it contained a
bit index.

Second, when a suitable bit index has been found that can contain the
required number of pages, the corresponding mmap[i] entry must be
checked to see if it can fit them all. The memory is not contigous!