Enabled preemptive multitasking
Preemptive multitasking is now enabled by scheduling on timer
interrupts.

Added the flag -fomit-frame-pointer to make gcc NOT use ebp for setting a new
stackframe on entry to a function (an optimization). The scheduler did
not work due to this, since it relied on ebp being used instead of esp.
This was fixed (hence the introduciton of run_process_in_user_mode and
run_process_in_kernel_mode).